relocate top_table so that preprocessor_utils and preprocessor monocle selection genes use the same one from utils
